chore: organize files in send-video module
