+++ packages/legacy/core/App/navigators/RootStack.tsx
@@ -20,7 +20,7 @@ import { DispatchAction } from '../contexts/reducers/store'
 import { useStore } from '../contexts/store'
 import { useTheme } from '../contexts/theme'
 import { useDeepLinks } from '../hooks/deep-links'
-import HistoryStack from '../modules/history/navigation/HistoryStack'
+import HistoryStack from '../modules/history/navigators/HistoryStack'
 import Chat from '../screens/Chat'
 import { BifoldError } from '../types/error'
 import { AuthenticateStackParams, Screens, Stacks, TabStacks } from '../types/navigators'
@@ -33,7 +33,7 @@ import DeliveryStack from './DeliveryStack'
 import NotificationStack from './NotificationStack'
 import ProofRequestStack from './ProofRequestStack'
 import SettingStack from './SettingStack'
-import SendVideoStack from '../modules/send-video/SendVideoStack'
+import SendVideoStack from '../modules/send-video/navigators/SendVideoStack'
 import TabStack from './TabStack'
 import { useDefaultStackOptions } from './defaultStackOptions'
